# openapi/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Query clarifier using NLP (Golang)",
    "description": "This is the DB microservice it contains two kinds of endpoints \n1.Query and 2.user\nIt stores the questions and answers and also updates the count as well\nand it is also gives the data of most frequently asked Questions this will be performed by the Database and gives the array of output ",
    "version": "1.0.1"
  },
  "paths": {
...
    "/chatbot/{Question}": {
      "get": {
        "tags": [
          "chatbot"
        ],
        "summary": "gets the data of the Question and answer present in the database",
        "description": "gets the data of the Question and answer present in the database",
        "operationId": "getbyQuestion",
        "parameters": [
          {
            "name": "Question",
            "in": "path",
            "description": "question of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "chatbot"
        ],
        "summary": "Will update the count in the db",
        "description": "Increases the count field",
        "operationId": "updatecount",
        "parameters": [
          {
            "name": "Question",
            "in": "path",
            "description": "question of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Username and Password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 10
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DatabaseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "32f7e7b3-4495-438c-a8a3-2c2f53fd5d76"
          },
          "question": {
            "type": "string",
            "example": "What is the college code of 2nd campus?"
          },
          "solution": {
            "type": "string",
            "example": "The code is :ACET"
          },
          "count": {
            "type": "integer",
            "example": 4
          }
        }
      },
...
    },
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      }
    }
  }
}